+++ src/Opaleye/Constant.hs
@@ -1,4 +1,0 @@
-module Opaleye.Constant {-# DEPRECATED "Use \"Opaleye.ToFields\" instead.  Will be removed in version 0.8." #-}
-  (module Opaleye.Internal.Constant) where
-
-import Opaleye.Internal.Constant