+++ src/Opaleye/Internal/Join.hs
@@ -135,35 +135,3 @@ instance Profunctor NullMaker where
 instance PP.ProductProfunctor NullMaker where
   purePP = pure
   (****) = (<*>)
-
---
-
-{-# DEPRECATED Nulled "Will be removed in version 0.8" #-}
-data Nulled
-
-type instance TF.IMap Nulled TF.OT     = TF.NullsT
-type instance TF.IMap Nulled TF.NullsT = TF.NullsT
-
--- It's quite unfortunate that we have to write these out by hand
--- until we probably do nullability as a distinction between
---
--- Column (Nullable a)
--- Column (NonNullable a)
-
-type instance Map.Map Nulled (Column (Nullable a)) = Column (Nullable a)
-
-type instance Map.Map Nulled (Column T.PGInt4) = Column (Nullable T.PGInt4)
-type instance Map.Map Nulled (Column T.PGInt8) = Column (Nullable T.PGInt8)
-type instance Map.Map Nulled (Column T.PGText) = Column (Nullable T.PGText)
-type instance Map.Map Nulled (Column T.PGFloat8) = Column (Nullable T.PGFloat8)
-type instance Map.Map Nulled (Column T.PGBool) = Column (Nullable T.PGBool)
-type instance Map.Map Nulled (Column T.PGUuid) = Column (Nullable T.PGUuid)
-type instance Map.Map Nulled (Column T.PGBytea) = Column (Nullable T.PGBytea)
-type instance Map.Map Nulled (Column T.PGText) = Column (Nullable T.PGText)
-type instance Map.Map Nulled (Column T.PGDate) = Column (Nullable T.PGDate)
-type instance Map.Map Nulled (Column T.PGTimestamp) = Column (Nullable T.PGTimestamp)
-type instance Map.Map Nulled (Column T.PGTimestamptz) = Column (Nullable T.PGTimestamptz)
-type instance Map.Map Nulled (Column T.PGTime) = Column (Nullable T.PGTime)
-type instance Map.Map Nulled (Column T.PGCitext) = Column (Nullable T.PGCitext)
-type instance Map.Map Nulled (Column T.PGJson) = Column (Nullable T.PGJson)
-type instance Map.Map Nulled (Column T.PGJsonb) = Column (Nullable T.PGJsonb)