+++ src/Opaleye/Map.hs
@@ -1,4 +1,0 @@
-module Opaleye.Map {-# DEPRECATED "This module will be removed in 0.8." #-}
-  (module Opaleye.Internal.Map) where
-
-import Opaleye.Internal.Map