+++ src/Opaleye/PGTypes.hs
@@ -1,4 +1,0 @@
-module Opaleye.PGTypes {-# DEPRECATED "Use \"Opaleye.SqlTypes\" instead." #-}
-  (module Opaleye.Internal.PGTypesExternal) where
-
-import Opaleye.Internal.PGTypesExternal