+++ src/Opaleye/QueryArr.hs
@@ -1,3 +1,0 @@
-module Opaleye.QueryArr {-# DEPRECATED "Use \"Opaleye.Select\" instead.  This module will be removed in 0.8." #-} (Query, QueryArr) where
-
-import           Opaleye.Internal.QueryArr (QueryArr, Query)