+++ src/Opaleye/RunQuery.hs
@@ -1,4 +1,0 @@
-module Opaleye.RunQuery {-# DEPRECATED "Use \"Opaleye.RunSelect\" instead." #-}
-  (module Opaleye.Internal.RunQueryExternal) where
-
-import Opaleye.Internal.RunQueryExternal