+++ src/Opaleye/Sql.hs
@@ -7,11 +7,6 @@ module Opaleye.Sql (
   -- * Explicit versions
   showSqlExplicit,
   showSqlUnoptExplicit,
-  -- * Deprecated functions
-  showSqlForPostgres,
-  showSqlForPostgresUnopt,
-  showSqlForPostgresExplicit,
-  showSqlForPostgresUnoptExplicit,
   ) where
 
 import qualified Opaleye.Internal.Unpackspec as U
@@ -59,21 +54,3 @@ showSqlExplicit = Pr.formatAndShowSQL
 
 showSqlUnoptExplicit :: U.Unpackspec fields b -> S.Select fields -> Maybe String
 showSqlUnoptExplicit = Pr.formatAndShowSQL .: Q.runQueryArrUnpack
-
-{-# DEPRECATED showSqlForPostgres "Will be removed in version 0.8.  Use 'showSql' instead." #-}
-showSqlForPostgres :: forall columns . D.Default U.Unpackspec columns columns =>
-                      S.Select columns -> Maybe String
-showSqlForPostgres = showSql
-
-{-# DEPRECATED showSqlForPostgresUnopt "Will be removed in version 0.8.  Use 'showSqlUnopt' instead." #-}
-showSqlForPostgresUnopt :: forall columns . D.Default U.Unpackspec columns columns =>
-                           S.Select columns -> Maybe String
-showSqlForPostgresUnopt = showSqlUnopt
-
-{-# DEPRECATED showSqlForPostgresExplicit "Will be removed in version 0.8.  Use 'showSqlExplicit' instead." #-}
-showSqlForPostgresExplicit :: U.Unpackspec columns b -> S.Select columns -> Maybe String
-showSqlForPostgresExplicit = showSqlExplicit
-
-{-# DEPRECATED showSqlForPostgresUnoptExplicit "Will be removed in version 0.8.  Use 'showSqlUnoptExplicit' instead." #-}
-showSqlForPostgresUnoptExplicit :: U.Unpackspec columns b -> S.Select columns -> Maybe String
-showSqlForPostgresUnoptExplicit = showSqlUnoptExplicit