+++ src/Opaleye/TypeFamilies.hs
@@ -17,7 +17,6 @@ module Opaleye.TypeFamilies
   , TF.Opt
   , TF.Req
   , TF.Nulls
-  , TF.TableField
   ) where
 
 import Opaleye.Internal.TypeFamilies as TF